requirements.txt
requests
bs4
re
datetime 
pandas
matplotlib.pyplot